removed `path_generator.py`
this was adding unnecessary complexity to the pipeline, in order to make a meaningful pathing class, the folder structure of qualisys, pupil, and unity data would need to be more standardized across data types so that it would be worth using a function to build paths as opposed to setting it up more manually (like I'm doing in lines 12-22 of `main.py`).

I didn't want to spend any more time on this than I already had. On to next steps!